Fix error in Theano when using GPU and TF (#19)
The configuration of the GPU device in Theano is set if USE_TF
is false and USE_GPU is true. Since USE_TF is only set to True
when using AWS EC2, it's now also set as a parameter in the
run_experiment_list function.

To avoid the conflict with Theano when using TF and GPU, set
both use_tf and use_gpu to True when calling
run_experiment_list.

to enable and disable GPU with TensorFlow, an if statement
was inserted to make the GPU invisible when use_gpu=False.

Also, an environment variable for use_tf for non-local modes
such as local_docker or EC2 was added if it's required to
switch between Theano and TensorFlow.

Finally, the corresponding documentation was added in
experiments.rst.